feat(#150): add license for all the classes
