+++ src/main/java/org/eolang/opeo/decompilation/DecompilerMachine.java
@@ -24,59 +24,19 @@
 package org.eolang.opeo.decompilation;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.Deque;
 import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
 import org.cactoos.list.ListOf;
-import org.cactoos.map.MapEntry;
-import org.cactoos.map.MapOf;
 import org.eolang.opeo.Instruction;
-import org.eolang.opeo.ast.Add;
-import org.eolang.opeo.ast.ArrayConstructor;
-import org.eolang.opeo.ast.AstNode;
-import org.eolang.opeo.ast.Attributes;
-import org.eolang.opeo.ast.ClassField;
-import org.eolang.opeo.ast.Constructor;
-import org.eolang.opeo.ast.Field;
-import org.eolang.opeo.ast.FieldAssignment;
-import org.eolang.opeo.ast.FieldRetrieval;
-import org.eolang.opeo.ast.Invocation;
-import org.eolang.opeo.ast.Label;
-import org.eolang.opeo.ast.Literal;
-import org.eolang.opeo.ast.LocalVariable;
-import org.eolang.opeo.ast.Mul;
-import org.eolang.opeo.ast.Opcode;
-import org.eolang.opeo.ast.Reference;
 import org.eolang.opeo.ast.Root;
-import org.eolang.opeo.ast.StaticInvocation;
-import org.eolang.opeo.ast.StoreArray;
-import org.eolang.opeo.ast.Substraction;
-import org.eolang.opeo.ast.Super;
-import org.eolang.opeo.ast.Typed;
-import org.eolang.opeo.ast.VariableAssignment;
-import org.eolang.opeo.jeo.JeoLabel;
-import org.objectweb.asm.Opcodes;
-import org.objectweb.asm.Type;
+import org.eolang.opeo.decompilation.handlers.RouterHandler;
 import org.xembly.Directive;
 
 /**
  * Decompiler machine.
  * @since 0.1
- * @todo #136:90min Refactor DecompilerMachine to simplify the code.
- *  The class is too complex and has too many responsibilities. It should be
- *  refactored to have a single responsibility. The class should be split into
- *  smaller classes, each with a single responsibility.
- * @checkstyle ClassFanOutComplexityCheck (500 lines)
  */
 public final class DecompilerMachine {
-
-    /**
-     * Output Stack.
-     */
-    private final Deque<AstNode> stack;
 
     /**
      * Local variables.
@@ -84,27 +44,22 @@ public final class DecompilerMachine {
     private final LocalVariables locals;
 
     /**
-     * Instruction handlers.
+     * Handler that redirects instructions.
      */
-    private final Map<Integer, InstructionHandler> handlers;
-
-    /**
-     * Arguments provided to decompiler.
-     */
-    private final Map<String, String> arguments;
+    private final RouterHandler router;
 
     /**
      * Constructor.
      */
-    public DecompilerMachine() {
-        this(new HashMap<>());
+    DecompilerMachine() {
+        this(new HashMap<>(0));
     }
 
     /**
      * Constructor.
      * @param args Arguments provided to decompiler.
      */
-    public DecompilerMachine(final Map<String, String> args) {
+    DecompilerMachine(final Map<String, String> args) {
         this(new LocalVariables(), args);
     }
 
@@ -114,46 +69,9 @@ public final class DecompilerMachine {
      * @param arguments Arguments provided to decompiler.
      */
     public DecompilerMachine(final LocalVariables locals, final Map<String, String> arguments) {
-        this.stack = new LinkedList<>();
         this.locals = locals;
-        this.arguments = arguments;
-        this.handlers = new MapOf<>(
-            new MapEntry<>(Opcodes.ICONST_0, new IconstHandler()),
-            new MapEntry<>(Opcodes.ICONST_1, new IconstHandler()),
-            new MapEntry<>(Opcodes.ICONST_2, new IconstHandler()),
-            new MapEntry<>(Opcodes.ICONST_3, new IconstHandler()),
-            new MapEntry<>(Opcodes.ICONST_4, new IconstHandler()),
-            new MapEntry<>(Opcodes.ICONST_5, new IconstHandler()),
-            new MapEntry<>(Opcodes.IADD, new AddHandler()),
-            new MapEntry<>(Opcodes.IADD, new AddHandler()),
-            new MapEntry<>(Opcodes.ISUB, new SubstractionHandler()),
-            new MapEntry<>(Opcodes.LSUB, new SubstractionHandler()),
-            new MapEntry<>(Opcodes.IMUL, new MulHandler()),
-            new MapEntry<>(Opcodes.ILOAD, new LoadHandler(Type.INT_TYPE)),
-            new MapEntry<>(Opcodes.LLOAD, new LoadHandler(Type.LONG_TYPE)),
-            new MapEntry<>(Opcodes.FLOAD, new LoadHandler(Type.FLOAT_TYPE)),
-            new MapEntry<>(Opcodes.DLOAD, new LoadHandler(Type.DOUBLE_TYPE)),
-            new MapEntry<>(Opcodes.ALOAD, new LoadHandler(Type.getType(Object.class))),
-            new MapEntry<>(Opcodes.ISTORE, new StoreHandler(Type.INT_TYPE)),
-            new MapEntry<>(Opcodes.LSTORE, new StoreHandler(Type.LONG_TYPE)),
-            new MapEntry<>(Opcodes.FSTORE, new StoreHandler(Type.FLOAT_TYPE)),
-            new MapEntry<>(Opcodes.DSTORE, new StoreHandler(Type.DOUBLE_TYPE)),
-            new MapEntry<>(Opcodes.ASTORE, new StoreHandler(Type.getType(Object.class))),
-            new MapEntry<>(Opcodes.AASTORE, new StoreToArrayHandler()),
-            new MapEntry<>(Opcodes.ANEWARRAY, new NewArrayHandler()),
-            new MapEntry<>(Opcodes.NEW, new NewHandler()),
-            new MapEntry<>(Opcodes.DUP, new DupHandler()),
-            new MapEntry<>(Opcodes.BIPUSH, new BipushHandler()),
-            new MapEntry<>(Opcodes.INVOKESPECIAL, new InvokespecialHandler()),
-            new MapEntry<>(Opcodes.INVOKEVIRTUAL, new InvokevirtualHandler()),
-            new MapEntry<>(Opcodes.INVOKESTATIC, new InvokestaticHander()),
-            new MapEntry<>(Opcodes.GETFIELD, new GetFieldHandler()),
-            new MapEntry<>(Opcodes.PUTFIELD, new PutFieldHnadler()),
-            new MapEntry<>(Opcodes.GETSTATIC, new GetStaticHnadler()),
-            new MapEntry<>(Opcodes.LDC, new LdcHandler()),
-            new MapEntry<>(Opcodes.POP, new PopHandler()),
-            new MapEntry<>(Opcodes.RETURN, new ReturnHandler()),
-            new MapEntry<>(JeoLabel.LABEL_OPCODE, new LabelHandler())
+        this.router = new RouterHandler(
+            arguments.getOrDefault("counting", "true").equals("true")
         );
     }
 
@@ -163,544 +81,10 @@ public final class DecompilerMachine {
      * @return Decompiled instructions.
      */
     public Iterable<Directive> decompile(final Instruction... instructions) {
-        MachineState state = new MachineState(this.locals);
+        final MachineState state = new MachineState(this.locals);
         Arrays.stream(instructions)
-            .forEach(inst -> this.handler(inst.opcode()).handle(state.next(inst)));
+            .forEach(inst -> this.router.handle(state.next(inst)));
         return new Root(new ListOf<>(state.stack().descendingIterator())).toXmir();
-    }
-
-    /**
-     * Do we add number to opcode name or not?
-     * if true then we add number to opcode name:
-     *  RETURN -> RETURN-1
-     * if false then we do not add number to opcode name:
-     *  RETURN -> RETURN
-     * @return Opcodes counting.
-     */
-    private boolean counting() {
-        return this.arguments.getOrDefault("counting", "true").equals("true");
-    }
-
-    /**
-     * Get instruction handler.
-     * @param opcode Instruction opcode.
-     * @return Instruction handler.
-     */
-    private InstructionHandler handler(final int opcode) {
-        return this.handlers.getOrDefault(opcode, new UnimplementedHandler());
-    }
-
-    /**
-     * Instruction handler.
-     * @since 0.1
-     */
-    private interface InstructionHandler {
-
-        /**
-         * Handle instruction.
-         * @param state Current instruction to handle together with operand stack and variables.
-         */
-        void handle(MachineState state);
-
-    }
-
-    /**
-     * Aload instruction handler.
-     * @since 0.1
-     */
-    private final class LoadHandler implements InstructionHandler {
-
-        /**
-         * Type of the variable.
-         */
-        private final Type type;
-
-        /**
-         * Constructor.
-         * @param type Type of the variable.
-         */
-        private LoadHandler(final Type type) {
-            this.type = type;
-        }
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(
-                state.variable(
-                    (Integer) state.operand(0),
-                    this.type
-                )
-            );
-        }
-    }
-
-    /**
-     * Store instruction handler.
-     * @since 0.1
-     */
-    private final class StoreHandler implements InstructionHandler {
-
-        /**
-         * Type of the variable.
-         */
-        private final Type type;
-
-        /**
-         * Constructor.
-         * @param type Type of the variable.
-         */
-        private StoreHandler(final Type type) {
-            this.type = type;
-        }
-
-        @Override
-        public void handle(final MachineState state) {
-            final OperandStack stack = state.stack();
-            final AstNode value = stack.pop();
-            stack.push(
-                new VariableAssignment(
-                    (LocalVariable) state.variable((Integer) state.operand(0), this.infer(value)),
-                    value
-                )
-            );
-        }
-
-        /**
-         * Infer type of the variable.
-         * @param value Value to infer type from.
-         * @return Inferred type.
-         */
-        private Type infer(final AstNode value) {
-            final Type result;
-            if (value instanceof Typed) {
-                result = ((Typed) value).type();
-            } else {
-                result = this.type;
-            }
-            return result;
-        }
-
-    }
-
-    /**
-     * Store to array instruction handler.
-     * Store a reference in an array
-     * Opcodes: aastore
-     * Stack [before]->[after]: "arrayref, index, value →"
-     * @since 0.1
-     */
-    private class StoreToArrayHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            final AstNode value = state.stack().pop();
-            final AstNode index = state.stack().pop();
-            final Reference array = (Reference) state.stack().pop();
-            array.link(new StoreArray(array.object(), index, value));
-        }
-    }
-
-    /**
-     * New array instruction handler.
-     * @since 0.1
-     */
-    private class NewArrayHandler implements InstructionHandler {
-        @Override
-        public void handle(final MachineState state) {
-            final String type = (String) state.operand(0);
-            final OperandStack stack = state.stack();
-            final AstNode size = stack.pop();
-            stack.push(new Reference(new ArrayConstructor(size, type)));
-        }
-
-    }
-
-    /**
-     * New instruction handler.
-     * @since 0.1
-     */
-    private class NewHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(new Reference());
-        }
-
-    }
-
-    /**
-     * Dup instruction handler.
-     * @since 0.1
-     */
-    private class DupHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().dup();
-        }
-
-    }
-
-    /**
-     * Getfield instruction handler.
-     * @since 0.1
-     */
-    private class GetFieldHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            final String owner = (String) state.operand(0);
-            final String name = (String) state.operand(1);
-            final String descriptor = (String) state.operand(2);
-            state.stack().push(
-                new FieldRetrieval(
-                    state.stack().pop(),
-                    new Attributes()
-                        .name(name)
-                        .descriptor(descriptor)
-                        .owner(owner)
-                        .type("field")
-                )
-            );
-        }
-
-    }
-
-    /**
-     * Putfield instruction handler.
-     * Stack [before]->[after]:  "objectref, value →"
-     * @since 0.1
-     */
-    private class PutFieldHnadler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            final AstNode value = state.stack().pop();
-            final String name = (String) state.operand(1);
-            final String owner = (String) state.operand(0);
-            final String descriptor = (String) state.operand(2);
-            state.stack().push(
-                new FieldAssignment(
-                    new Field(
-                        state.stack().pop(),
-                        new Attributes()
-                            .name(name)
-                            .owner(owner)
-                            .descriptor(descriptor)
-                    ),
-                    value
-                )
-            );
-        }
-
-    }
-
-    /**
-     * Bipush instruction handler.
-     * @since 0.1
-     */
-    private class BipushHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(new Literal(state.operand(0)));
-        }
-
-    }
-
-    /**
-     * Pop instruction handler.
-     * @since 0.1
-     */
-    private class PopHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            // We do nothing here to keep the stack contains previous computations.
-        }
-
-    }
-
-    /**
-     * Return instruction handler.
-     * @since 0.1
-     */
-    private class ReturnHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(
-                new Opcode(
-                    state.instruction().opcode(),
-                    state.instruction().operands(),
-                    DecompilerMachine.this.counting()
-                )
-            );
-        }
-
-    }
-
-    /**
-     * Invokespecial instruction handler.
-     * @since 0.1
-     */
-    private class InvokespecialHandler implements InstructionHandler {
-        @Override
-        public void handle(final MachineState state) {
-            if (!state.operand(1).equals("<init>")) {
-                throw new UnsupportedOperationException(
-                    String.format("Instruction %s is not supported yet", state)
-                );
-            }
-            final String descriptor = (String) state.operand(2);
-            final List<AstNode> args = state.stack().pop(
-                Type.getArgumentCount(descriptor)
-            );
-            final String target = (String) state.operand(0);
-            //@checkstyle MethodBodyCommentsCheck (10 lines)
-            // @todo #76:90min Target might not be an Object.
-            //  Here we just compare with object, but if the current class has a parent, the
-            //  target might not be an Object. We should compare with the current class name
-            //  instead. Moreover, we have to pass the 'target' as an argument to the
-            //  constructor of the 'Super' class somehow.
-            if ("java/lang/Object".equals(target)) {
-                state.stack().push(
-                    new Super(state.stack().pop(), args, descriptor)
-                );
-            } else {
-                ((Reference) state.stack().pop())
-                    .link(new Constructor(target, new Attributes().descriptor(descriptor), args));
-            }
-        }
-
-    }
-
-    /**
-     * Invokevirtual instruction handler.
-     * @since 0.1
-     */
-    private class InvokevirtualHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            final String owner = (String) state.operand(0);
-            final String method = (String) state.operand(1);
-            final String descriptor = (String) state.operand(2);
-            final List<AstNode> args = state.stack().pop(
-                Type.getArgumentCount(descriptor)
-            );
-            Collections.reverse(args);
-            final AstNode source = state.stack().pop();
-            state.stack().push(
-                new Invocation(
-                    source,
-                    new Attributes().name(method).descriptor(descriptor).owner(owner),
-                    args
-                )
-            );
-        }
-    }
-
-    /**
-     * Invokestatic instruction handler.
-     * @since 0.1
-     */
-    private class InvokestaticHander implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            final String owner = (String) state.operand(0);
-            final String method = (String) state.operand(1);
-            final String descriptor = (String) state.operand(2);
-            final List<AstNode> args = state.stack().pop(
-                Type.getArgumentCount(descriptor)
-            );
-            state.stack().push(
-                new StaticInvocation(owner, method, descriptor, args)
-            );
-        }
-    }
-
-    /**
-     * Ldc instruction handler.
-     * @since 0.1
-     */
-    private class LdcHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(new Literal(state.operand(0)));
-        }
-
-    }
-
-    /**
-     * Unimplemented instruction handler.
-     * @since 0.1
-     */
-    private class UnimplementedHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(
-                new Opcode(
-                    state.instruction().opcode(),
-                    state.instruction().operands(),
-                    DecompilerMachine.this.counting()
-                )
-            );
-        }
-
-    }
-
-    /**
-     * Add instruction handler.
-     * @since 0.1
-     */
-    private class AddHandler implements InstructionHandler {
-        @Override
-        public void handle(final MachineState state) {
-            if (state.instruction().opcode() == Opcodes.IADD) {
-                final AstNode right = state.stack().pop();
-                final AstNode left = state.stack().pop();
-                state.stack().push(new Add(left, right));
-            } else if (state.instruction().opcode() == Opcodes.LADD) {
-                final AstNode right = state.stack().pop();
-                final AstNode left = state.stack().pop();
-                state.stack().push(
-                    new Add(left, right, new Attributes().type("long"))
-                );
-            } else {
-                state.stack().push(
-                    new Opcode(
-                        state.instruction().opcode(),
-                        state.instruction().operands(),
-                        DecompilerMachine.this.counting()
-                    )
-                );
-            }
-        }
-
-    }
-
-    /**
-     * Substraction instruction handler.
-     * @since 0.1
-     */
-    private class SubstractionHandler implements InstructionHandler {
-        @Override
-        public void handle(final MachineState state) {
-            if (state.instruction().opcode() == Opcodes.ISUB) {
-                final AstNode right = state.stack().pop();
-                final AstNode left = state.stack().pop();
-                state.stack().push(
-                    new Substraction(left, right, new Attributes().type("int"))
-                );
-            } else if (state.instruction().opcode() == Opcodes.LSUB) {
-                final AstNode right = state.stack().pop();
-                final AstNode left = state.stack().pop();
-                state.stack().push(
-                    new Substraction(left, right, new Attributes().type("long"))
-                );
-            } else {
-                state.stack().push(
-                    new Opcode(
-                        state.instruction().opcode(),
-                        state.instruction().operands(),
-                        DecompilerMachine.this.counting()
-                    )
-                );
-            }
-        }
-
-    }
-
-    /**
-     * Mul instruction handler.
-     * @since 0.1
-     */
-    private class MulHandler implements InstructionHandler {
-        @Override
-        public void handle(final MachineState state) {
-            if (state.instruction().opcode() == Opcodes.IMUL) {
-                final AstNode right = state.stack().pop();
-                final AstNode left = state.stack().pop();
-                state.stack().push(new Mul(left, right));
-            } else {
-                state.stack().push(
-                    new Opcode(
-                        state.instruction().opcode(),
-                        state.instruction().operands(),
-                        DecompilerMachine.this.counting()
-                    )
-                );
-            }
-        }
-
-    }
-
-    /**
-     * Iconst instruction handler.
-     * @since 0.1
-     */
-    private class IconstHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            switch (state.instruction().opcode()) {
-                case Opcodes.ICONST_0:
-                    state.stack().push(new Literal(0));
-                    break;
-                case Opcodes.ICONST_1:
-                    state.stack().push(new Literal(1));
-                    break;
-                case Opcodes.ICONST_2:
-                    state.stack().push(new Literal(2));
-                    break;
-                case Opcodes.ICONST_3:
-                    state.stack().push(new Literal(3));
-                    break;
-                case Opcodes.ICONST_4:
-                    state.stack().push(new Literal(4));
-                    break;
-                case Opcodes.ICONST_5:
-                    state.stack().push(new Literal(5));
-                    break;
-                default:
-                    throw new UnsupportedOperationException(
-                        String.format("Instruction %s is not supported yet", state.instruction())
-                    );
-            }
-        }
-
-    }
-
-    /**
-     * Label instruction handler.
-     * @since 0.1
-     */
-    private class LabelHandler implements InstructionHandler {
-
-        @Override
-        public void handle(final MachineState state) {
-            state.stack().push(
-                new Label(new Literal(state.operand(0)))
-            );
-        }
-    }
-
-    /**
-     * Getstatic instruction handler.
-     * @since 0.1
-     */
-    private class GetStaticHnadler implements InstructionHandler {
-        @Override
-        public void handle(final MachineState state) {
-            final String klass = (String) state.operand(0);
-            final String method = (String) state.operand(1);
-            final String descriptor = (String) state.operand(2);
-            state.stack().push(new ClassField(klass, method, descriptor));
-        }
     }
 }
