+++ src/main/java/org/eolang/opeo/decompilation/handlers/GetStaticHnadler.java
@@ -1,0 +1,42 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016-2023 Objectionary.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package org.eolang.opeo.decompilation.handlers;
+
+import org.eolang.opeo.ast.ClassField;
+import org.eolang.opeo.decompilation.InstructionHandler;
+import org.eolang.opeo.decompilation.MachineState;
+
+/**
+ * Getstatic instruction handler.
+ * @since 0.1
+ */
+public class GetStaticHnadler implements InstructionHandler {
+    @Override
+    public void handle(final MachineState state) {
+        final String klass = (String) state.operand(0);
+        final String method = (String) state.operand(1);
+        final String descriptor = (String) state.operand(2);
+        state.stack().push(new ClassField(klass, method, descriptor));
+    }
+}
