+++ src/main/java/org/eolang/opeo/decompilation/handlers/StoreHandler.java
@@ -1,0 +1,81 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016-2023 Objectionary.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package org.eolang.opeo.decompilation.handlers;
+
+import org.eolang.opeo.ast.AstNode;
+import org.eolang.opeo.ast.LocalVariable;
+import org.eolang.opeo.ast.Typed;
+import org.eolang.opeo.ast.VariableAssignment;
+import org.eolang.opeo.decompilation.InstructionHandler;
+import org.eolang.opeo.decompilation.MachineState;
+import org.eolang.opeo.decompilation.OperandStack;
+import org.objectweb.asm.Type;
+
+/**
+ * Store instruction handler.
+ * @since 0.1
+ */
+public final class StoreHandler implements InstructionHandler {
+
+    /**
+     * Type of the variable.
+     */
+    private final Type type;
+
+    /**
+     * Constructor.
+     * @param type Type of the variable.
+     */
+    public StoreHandler(final Type type) {
+        this.type = type;
+    }
+
+    @Override
+    public void handle(final MachineState state) {
+        final OperandStack stack = state.stack();
+        final AstNode value = stack.pop();
+        stack.push(
+            new VariableAssignment(
+                (LocalVariable) state.variable((Integer) state.operand(0), this.infer(value)),
+                value
+            )
+        );
+    }
+
+    /**
+     * Infer type of the variable.
+     * @param value Value to infer type from.
+     * @return Inferred type.
+     */
+    private Type infer(final AstNode value) {
+        final Type result;
+        if (value instanceof Typed) {
+            result = ((Typed) value).type();
+        } else {
+            result = this.type;
+        }
+        return result;
+    }
+
+}
